[discretizers] move convert_to_numpy_list_vector_array to pymor.discretizers.builtin.list
