+++ src/pymordemos/thermalblock.py
@@ -229,7 +229,7 @@ def parse_arguments(args):
 def discretize_pymor(xblocks, yblocks, grid_num_intervals, use_list_vector_array):
     from pymor.analyticalproblems.thermalblock import thermal_block_problem
     from pymor.discretizers.builtin import discretize_stationary_cg
-    from pymor.playground.discretizers.numpylistvectorarray import convert_to_numpy_list_vector_array
+    from pymor.discretizers.builtin.list import convert_to_numpy_list_vector_array
 
     print('Discretize ...')
     # setup analytical problem
